+++ lib/transpec/rewriter.rb
@@ -152,12 +152,16 @@ module Transpec
     def need_to_modify_expectation_syntax_configuration?(rspec_configure)
       return false unless @configuration.convert_to_expect_to_matcher?
       rspec_configure.expectation_syntaxes == [:should]
+    rescue Syntax::RSpecConfigure::UnknownSyntaxError
+      false
     end
 
     def need_to_modify_mock_syntax_configuration?(rspec_configure)
       return false if !@configuration.convert_to_expect_to_receive? &&
                       !@configuration.convert_to_allow_to_receive?
       rspec_configure.mock_syntaxes == [:should]
+    rescue Syntax::RSpecConfigure::UnknownSyntaxError
+      false
     end
 
     class OverlappedRewriteError < StandardError; end